+++ csrc/hybrid_ep/extension/permute.cuh
@@ -22,7 +22,7 @@ struct PermuteArgs {
   int hidden_size;
   int scales_per_token; // Now is hidden_size/128
   torch::Tensor num_dispatched_token_tensor; // We assume it is only valid on GPU
-  int num_permuted_token;
+  int64_t num_permuted_token;
   int num_ranks_per_node; // Probs dimension 0 = num_ranks_per_node * num_of_local_experts
   int num_of_local_experts;
   int pad_multiple;
@@ -77,11 +77,11 @@ struct UnpermuteArgs {
  permute_preprocessing(
      bool* routing_map,
      torch::Tensor num_dispatched_token_tensor,
-     int max_num_dispatched_tokens,
+     int64_t max_num_dispatched_tokens,
      int num_of_local_experts,
      int pad_multiple,
      int num_of_blocks,
-     int num_permuted_tokens,
+     int64_t num_permuted_tokens,
      bool non_blocking,
      cudaStream_t stream);
